Beginning tidy up. Changes to event loop monad.
The game wires now run on the non-IO GameM monad, which wraps the various states
of the game into the required interface.  At the moment this state is only the
random generator used for wackelkontaktM, but should soon become a full monad
transformer to tidy up the current disheveled state of the eventLoop function.